backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
aiosmtplib==3.0.2
aiohttp==3.11.9
pydantic==2.10.3
jinja2==3.1.4
email-validator==2.2.0
